refactor to better reuse drag drop table components
