+++ src/mcmfuserinterface/Controller.java
@@ -6,6 +6,8 @@
 package mcmfuserinterface;
 
 import model.MCMFModel;
+import model.Project;
+import model.Reader;
 
 /**
  *
@@ -15,4 +17,8 @@ public interface Controller {
     public MCMFModel getModel();
     public void refreshTable();
     public void refreshLowSelectedProjectList();
+    public int moveProject(Reader reader, Reader readerToRemoveFrom, Project projectToMove, Project projectToPlaceBefore);
+    public boolean moveProject(Reader readerToAdd, Reader readerToRemoveFrom, Project projectToMove);
+    public boolean addProjectToReader(Reader reader, Project projectToAdd);
+    public int addProjectToReader(Reader reader, Project projectToAdd, Project projectToAddBefore);
 }
