--- src/mcmfuserinterface/drag_drop_table/AssignmentsScrollPane.java
@@ -1,20 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package mcmfuserinterface.drag_drop_table;
-
-import mcmfuserinterface.Controller;
-
-/**
- *
- * @author Eimantas
- */
-public class AssignmentsScrollPane extends DroppableScrollPane{
-    
-    public AssignmentsScrollPane(final Controller controller){
-        super();
-    }
-    
-}
